pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "var-reasoning"
version = "0.1.0"
description = "Verification-Augmented Reasoning: Making LLMs reason better through code investigation and formal verification"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.0.0",
    "z3-solver>=4.12.0",
    "sympy>=1.12",
    "numpy>=1.26",
    "scipy>=1.12",
    "docker>=7.0",
    "datasets>=2.16",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "matplotlib>=3.8",
    "seaborn>=0.13",
    "tqdm>=4.66",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
